+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2023-03-21
+
 ### Removed
 
 - Remove faker from dev
@@ -29,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial setup
 
-[Unreleased]: https://github.com/sauljabin/changeloggh/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/sauljabin/changeloggh/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/sauljabin/changeloggh/compare/v0.2.0...v1.0.0
 [0.2.0]: https://github.com/sauljabin/changeloggh/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/sauljabin/changeloggh/releases/tag/v0.1.0